requirements.txt
requests
cpe
python-dateutil
